requirements.txt
gitpython==3.1.40        
cspell==7.3.1            
markdown==3.4.3          
rich==13.7.0             
pandas==2.1.4            
numpy==1.26.3            
pytest==7.4.3            
